lsp: Do not panic when showing a document to the user
