+++ tools/lsp/main.rs
@@ -458,14 +458,15 @@ async fn handle_notification(req: lsp_server::Notification, ctx: &Rc<Context>) -
                     crate::preview::notify_lsp_diagnostics(&ctx.server_notifier, uri, diagnostics);
                 }
                 M::ShowDocument { file, start_line, start_column, end_line, end_column } => {
-                    crate::preview::send_show_document_to_editor(
-                        &ctx.server_notifier,
-                        &file,
+                    send_show_document_to_editor(
+                        ctx.server_notifier.clone(),
+                        file,
                         start_line,
                         start_column,
                         end_line,
                         end_column,
-                    );
+                    )
+                    .await;
                 }
                 M::PreviewTypeChanged { is_external } => {
                     ctx.preview.set_use_external_previewer(is_external);
@@ -478,4 +479,28 @@ async fn handle_notification(req: lsp_server::Notification, ctx: &Rc<Context>) -
         _ => (),
     }
     Ok(())
+}
+
+pub async fn send_show_document_to_editor(
+    sender: ServerNotifier,
+    file: String,
+    start_line: u32,
+    start_column: u32,
+    end_line: u32,
+    end_column: u32,
+) {
+    let Some(params) = crate::preview::show_document_request_from_element_callback(
+        &file,
+        start_line,
+        start_column,
+        end_line,
+        end_column,
+    ) else {
+        return;
+    };
+    let Ok(fut) = sender.send_request::<lsp_types::request::ShowDocument>(params) else {
+        return;
+    };
+
+    let _ = fut.await;
 }
